completionProvider: Allow delayed inline completion
VSCode really wants inline completion providers to generate completion
items faster than we can realistically provide, given the LLM may
take the order of seconds to come up with a reply. In some cases
VSCode expects a response in less than 80ms.

In order to cope, this commit makes the inline completion provider
return null immediately, while working on the results in the background.

Once they come in, it tells VSCode to try again.

This commit also adds an onAccept listener to know when to clean up the
inline completion after the user has green lighted it.