core/autocomplete: Provide way to register completion accept handler
When the user hits tab to accept an inline completion the completion
provider that provided the completion gets notified via its accept
method.

There are other parts of the code that would ideally like to know
when the completion provider accept method is called.

This commit adds a VSCode-like onAccept function to register disposable
accept listeners.